Create transaction migration and update migrations for other tables so that ids are bigint data types

+++ db/migrate/20191001040803_create_items.rb
@@ -4,7 +4,7 @@ class CreateItems < ActiveRecord::Migration[5.1]
       t.string :name
       t.string :description
       t.float :unit_price
-      t.integer :merchant_id
+      t.bigint :merchant_id
       t.string :created_at
       t.text :updated_at
